requirements.txt
transformers
tokenizers
pandas
numpy
matplotlib
torch
streamlit-tags
sentencepiece